code/picode/visionAI/backend/requirements.txt
fastapi==0.109.2
termcolor==2.3.0
python-arango == 7.5.7
websockets==12.0
requests==2.31.0
uvicorn==0.27.1
pydantic-settings==2.2.1
opencv-python-headless
pyserial
pyzbar==0.1.9
aiofiles==23.2.1
python-multipart==0.0.9